examples/alphazero-2048/requirements.txt
pgx>=2.0.0
dm-haiku
mctx
optax
wandb
omegaconf
pydantic
imageio
cairosvg
imageio[ffmpeg]
